ShellPkg/acpi: Code cleanup to pass static code checker
Contributed-under: TianoCore Contribution Agreement 1.1

+++ ShellPkg/Library/UefiShellAcpiViewCommandLib/AcpiParser.c
@@ -297,7 +297,7 @@ DumpUint64 (
   Val = *(UINT32*)(Ptr + sizeof (UINT32));
 
   Val <<= 32;
-  Val |= *(UINT32*)Ptr;
+  Val |= (UINT64)*(UINT32*)Ptr;
 
   Print (Format, Val);
 }
